Wrap {filePath} in double quotes to handle filepaths with spaces
This change addresses the following output:

[error] No files matching the pattern were found: "/Users/benmaclaurin/Developer/gilleducation.ie/cli/create-block/../../blocks/block7/[Help".
[error] No files matching the pattern were found: "center].json".

The issue arises where filePath contains spaces. It will be interpreted as multiple arguments by the shell.

To solve this problem, we wrap the filePath in double quotes. This will ensure that the shell treats the entire file path as a single argument, even if it contains spaces.